Commit reordering: Avoid asking about the same conflict twice
It is fair to assume that the user never wants or expects
the sum of changes to be different after reordering them.
That would be an impure reordering,
which is not what the user is asking for.
Even if the conflicts are resolved by the user or git-rerere,
it is fair to assume such a result to be a failure.

This commit simply always upholds this invariant
for the practical benefit of not bothering the user
about a conflict with a known resolution.

I can hear critics say that this is an assumption,
and it's safer not to assume. The answer to that is plainly no:
There is only one correct solution to the second conflict given the first.
Not just in theory, but when you do this exercise day in and day out
and see that the second conflict is always a rearrangement of the first,
you want to spend your time on making the first one right.

Implements https://github.com/mystor/git-revise/issues/132