Fix missing no-change message in -ie mode
This warning is nice to see,
but only worked in `-i` and `-e` mode, not in `-ie` mode:

    (warning) no changes performed

The Step.message was initialized to None instead of the old value,
causing unedited messages in message-edit mode to compare unequal.

This commit removes the optionality and thereby the possibility
for the problem.
It can also be solved by interpreting None as the old value
or resetting the new value to None if it is indifferent.